Updated ExecutionManagers to run Attended and Server Jobs smoothly

+++ OpenBots.Service.Client/Manager/Execution/AttendedExecutionManager.cs
@@ -16,7 +16,7 @@ namespace OpenBots.Service.Client.Manager.Execution
 {
     public class AttendedExecutionManager
     {
-        public bool IsAttendedJobRunning { get; set; } = false;
+        //public bool IsAttendedJobRunning { get; set; } = false;
         public static AttendedExecutionManager Instance
         {
             get
@@ -35,42 +35,50 @@ namespace OpenBots.Service.Client.Manager.Execution
 
         public bool ExecuteTask(string projectPackagePath, ServerConnectionSettings settings)
         {
-            bool isSuccessful;
-            string projectDirectoryPath = string.Empty;
-            try
+            if (!ExecutionManager.Instance.IsEngineBusy)
             {
-                IsAttendedJobRunning = true;
-                string configFilePath;
-                string mainScriptFilePath = AutomationManager.GetMainScriptFilePath(projectPackagePath, out configFilePath);
-                projectDirectoryPath = Path.GetDirectoryName(mainScriptFilePath);
+                bool isSuccessful;
+                string projectDirectoryPath = string.Empty;
+                try
+                {
+                    ExecutionManager.Instance.SetEngineStatus(true);
+                    string configFilePath;
+                    string mainScriptFilePath = AutomationManager.GetMainScriptFilePath(projectPackagePath, out configFilePath);
+                    projectDirectoryPath = Path.GetDirectoryName(mainScriptFilePath);
 
-                NugetPackageManager.InstallProjectDependencies(configFilePath);
-                var assembliesList = NugetPackageManager.LoadPackageAssemblies(configFilePath);
+                    NugetPackageManager.InstallProjectDependencies(configFilePath);
+                    var assembliesList = NugetPackageManager.LoadPackageAssemblies(configFilePath);
 
-                RunAttendedAutomation(mainScriptFilePath, settings, assembliesList);
+                    RunAttendedAutomation(mainScriptFilePath, settings, assembliesList);
 
-                isSuccessful = true;
+                    isSuccessful = true;
+                }
+                catch (Exception)
+                {
+                    isSuccessful = false;
+                }
+                finally
+                {
+                    // Delete Project Directory
+                    if (Directory.Exists(projectDirectoryPath))
+                        Directory.Delete(projectDirectoryPath, true);
+
+                    ExecutionManager.Instance.SetEngineStatus(false);
+                }
+
+                return isSuccessful;
             }
-            catch (Exception)
-            {
-                isSuccessful = false;
-            }
-            finally
-            {
-                // Delete Project Directory
-                if(Directory.Exists(projectDirectoryPath))
-                    Directory.Delete(projectDirectoryPath, true);
-
-                IsAttendedJobRunning = false;
-            }
-
-            return isSuccessful;
+            return false;
         }
 
         private void RunAttendedAutomation(string mainScriptFilePath, ServerConnectionSettings settings, List<string> projectDependencies)
         {
             var executionParams = GetExecutionParams(mainScriptFilePath, settings, projectDependencies);
-            var userInfo = new Credential(settings.DNSHost, settings.UserName);
+            var userInfo = new MachineCredential
+            {
+                Domain = settings.DNSHost, 
+                UserName = settings.UserName
+            };
 
             var executorPath = Directory.GetFiles(AppDomain.CurrentDomain.BaseDirectory, "OpenBots.Executor.exe").FirstOrDefault();
             var cmdLine = $"\"{executorPath}\" \"{executionParams}\"";
